pyproject.toml
[project]
name = "commitiq"
version = "0.1.0"
description = "CLI tool that summarizes your git commits across repos as functional tasks using LLMs, with MCP server support"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Aritra Sadhukhan", email = "aritrasadhukhan430@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.3",
    "gitpython>=3.1.47",
    "litellm>=1.0.0",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
    "mcp>=1.0.0",
]

[project.scripts]
commitiq = "commitiq:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"
